tests/wpt/web-platform-tests/css/css-overflow/parsing/text-overflow-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Overflow: getComputedValue().textOverflow</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow/#propdef-text-overflow">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-overflow", 'clip');
test_computed_value("text-overflow", 'ellipsis');
</script>
</body>
</html>
